+++ docs/changes.rst
@@ -19,6 +19,7 @@ New features
 - Add :meth:`Schema.after_init_hook` method that gets called when schema is finished initializing.
 - Add ``load_key`` and ``dump_key`` parameters to :class:`fields.Field` to control the field names
   during serialization/deserialization of data.
+- Add :attr:`ValidationError.state` to allow storing state to an error that can later be accessed.
 
 Improvements
 ~~~~~~~~~~~~